Remove old test file
