--- .byebug_history
@@ -1,30 +1,0 @@
-exit
-tickets
-:tickets => tickets
-      }
-        :name => name
-        :error_timeout => error_timeout,
-        :error_threshold => error_threshold,
-        :success_threshold => success_threshold,
-        :tickets => tickets,
-{
-      }
-        name => name
-        error_timeout => error_timeout,
-        error_threshold => error_threshold,
-        success_threshold => success_threshold,
-        tickets => tickets,
-{
-tickets: tickets
-      }
-        name: name
-        error_timeout: error_timeout,
-        error_threshold: error_threshold,
-        success_threshold: success_threshold,
-        tickets: tickets,
-{
-c
-semian_options[:tickets]
-semian_options
-exit
-error
